requirements.txt
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gr==1.16.1
gunicorn==20.0.4
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
opencv-contrib-python==4.4.0.44
opencv-contrib-python-headless==4.4.0.44
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.10.1
Werkzeug==1.0.1
